+++ CHANGELOG.md
@@ -5,5 +5,7 @@
 ### Added
 
 - `csv2sheets`: reads bibliographic details from `papers.csv` and updates a Google Sheet
+- `csv2wordcloud`: reads abstracts and titles from `papers.csv` and generates wordclouds
 - `sheets2bib`: reads papers from a Google Sheet, looks up bibtex records, and writes to `references.bib`
 - `sheets2csv`: reads papers from a Google Sheet, looks up bibliographic details, and writes to `papers.csv`
+- `sheets2wordcloud`: reads abstracts and titles from a Google Sheet and generates wordclouds